fix: 2주차 미션#2 성적관리시스템에서 성적조회시 평균값 소수점 둘째자리까지 표시되게 수정

+++ week2_mission2/src/main/java/org/codesquad/shool/courses/CoursesList.java
@@ -24,9 +24,9 @@ public class CoursesList {
         return totalScore;
     }
 
-    public int averageScore() {
+    public double averageScore() {
         int totalScore = totalScore();
-        return totalScore / coursesList.size();
+        return totalScore / (double) coursesList.size();
     }
 
     public String infoFileWriteConverter() {
